pyproject.toml
[tool.poetry]
name = "business_assistant"
version = "0.1.0"
description = ""
authors = ["StepanTchynetskyiDS <stepan.tchynetskyi@deepsee.ai>"]

[tool.poetry.dependencies]
python = "3.10.4"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
flake8 = "^4.0.1"
SQLAlchemy = "^1.4.35"
alembic = "^1.7.7"
psycopg2 = "^2.9.3"
PyJWT = "^2.3.0"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
